+++ src/Analyzers/MSTest.Analyzers/PublicAPI.Shipped.txt
@@ -1,0 +1,1 @@
+﻿#nullable enable
