--- src/Platform/Microsoft.Testing.Platform/Helpers/SystemPolyfills.cs
@@ -1,569 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-#pragma warning disable SA1201 // Elements should appear in the correct order
-#pragma warning disable SA1313 // Parameter names should begin with lower-case letter
-#pragma warning disable SA1403 // File may only contain a single namespace
-#pragma warning disable SA1502 // Element should not be on a single line
-#pragma warning disable SA1623 // Property summary documentation should match accessors
-#pragma warning disable SA1642 // Constructor summary documentation should begin with standard text
-#pragma warning disable CS1574 // XML comment has cref attribute that could not be resolved
-
-#if !NETCOREAPP
-using System.ComponentModel;
-#if !NET462
-using System.Diagnostics.CodeAnalysis;
-using System.Runtime.CompilerServices;
-
-using Microsoft.Testing.Platform;
-#endif
-
-namespace System.Runtime.CompilerServices
-{
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    internal class IsExternalInit { }
-
-    [AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false, Inherited = false)]
-    internal sealed class CallerArgumentExpressionAttribute : Attribute
-    {
-        public CallerArgumentExpressionAttribute(string parameterName)
-        {
-            ParameterName = parameterName;
-        }
-
-        public string ParameterName { get; }
-    }
-}
-
-namespace System
-{
-    // https://stackoverflow.com/questions/263400/what-is-the-best-algorithm-for-overriding-gethashcode/263416#263416
-    internal struct HashCode
-    {
-        private int _hash;
-
-        public HashCode()
-        {
-            _hash = 17;
-        }
-
-        public void Add(string value)
-        {
-            // Overflow is fine, just wrap
-            unchecked
-            {
-                _hash = (_hash * 23) + (value?.GetHashCode() ?? 0);
-            }
-        }
-
-        public void Add(bool value)
-        {
-            // Overflow is fine, just wrap
-            unchecked
-            {
-                _hash = (_hash * 23) + value.GetHashCode();
-            }
-        }
-
-        public void Add(int value)
-        {
-            // Overflow is fine, just wrap
-            unchecked
-            {
-                _hash = (_hash * 23) + value.GetHashCode();
-            }
-        }
-
-        public readonly int ToHashCode()
-            => _hash;
-    }
-
-#if !NET462
-    /// <summary>Represent a type can be used to index a collection either from the start or the end.</summary>
-    /// <remarks>
-    /// Index is used by the C# compiler to support the new index syntax.
-    /// <code>
-    /// int[] someArray = new int[5] { 1, 2, 3, 4, 5 } ;
-    /// int lastElement = someArray[^1]; // lastElement = 5
-    /// </code>
-    /// </remarks>
-    [SuppressMessage("Performance", "CA1815:Override equals and operator equals on value types", Justification = "Matches implementation from dotnet/runtime")]
-    [SuppressMessage("Usage", "CA2231:Overload operator equals on overriding value type Equals", Justification = "Matches implementation from dotnet/runtime")]
-    internal readonly struct Index : IEquatable<Index>
-    {
-        private readonly int _value;
-
-        /// <summary>Initializes a new instance of the <see cref="Index"/> struct.Construct an Index using a value and indicating if the index is from the start or from the end.</summary>
-        /// <param name="value">The index value. it has to be zero or positive number.</param>
-        /// <param name="fromEnd">Indicating if the index is from the start or from the end.</param>
-        /// <remarks>
-        /// If the Index constructed from the end, index value 1 means pointing at the last element and index value 0 means pointing at beyond last element.
-        /// </remarks>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Index(int value, bool fromEnd = false)
-        {
-            if (value < 0)
-            {
-                throw new ArgumentOutOfRangeException(nameof(value), value, "Non-negative number required.");
-            }
-
-            _value = fromEnd ? ~value : value;
-        }
-
-        // The following private constructors mainly created for perf reason to avoid the checks
-        private Index(int value)
-        {
-            _value = value;
-        }
-
-        /// <summary>Gets create an Index pointing at first element.</summary>
-        public static Index Start => new(0);
-
-        /// <summary>Gets create an Index pointing at beyond last element.</summary>
-        public static Index End => new(~0);
-
-        /// <summary>Create an Index from the start at the position indicated by the value.</summary>
-        /// <param name="value">The index value from the start.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Index FromStart(int value)
-            => value < 0
-                ? throw new ArgumentOutOfRangeException(nameof(value), value, "Non-negative number required.")
-                : new Index(value);
-
-        /// <summary>Create an Index from the end at the position indicated by the value.</summary>
-        /// <param name="value">The index value from the end.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Index FromEnd(int value)
-            => value < 0
-                ? throw new ArgumentOutOfRangeException(nameof(value), value, "Non-negative number required.")
-                : new Index(~value);
-
-        /// <summary>Gets the index value.</summary>
-        public int Value => _value < 0 ? ~_value : _value;
-
-        /// <summary>Gets a value indicating whether indicates whether the index is from the start or the end.</summary>
-        public bool IsFromEnd => _value < 0;
-
-        /// <summary>Calculate the offset from the start using the giving collection length.</summary>
-        /// <param name="length">The length of the collection that the Index will be used with. length has to be a positive value.</param>
-        /// <remarks>
-        /// For performance reason, we don't validate the input length parameter and the returned offset value against negative values.
-        /// we don't validate either the returned offset is greater than the input length.
-        /// It is expected Index will be used with collections which always have non negative length/count. If the returned offset is negative and
-        /// then used to index a collection will get out of range exception which will be same affect as the validation.
-        /// </remarks>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public int GetOffset(int length)
-        {
-            int offset = _value;
-            if (IsFromEnd)
-            {
-                // offset = length - (~value)
-                // offset = length + (~(~value) + 1)
-                // offset = length + value + 1
-                offset += length + 1;
-            }
-
-            return offset;
-        }
-
-        /// <summary>Indicates whether the current Index object is equal to another object of the same type.</summary>
-        /// <param name="obj">An object to compare with this object.</param>
-        public override bool Equals(object? obj) => obj is Index index && _value == index._value;
-
-        /// <summary>Indicates whether the current Index object is equal to another Index object.</summary>
-        /// <param name="other">An object to compare with this object.</param>
-        public bool Equals(Index other) => _value == other._value;
-
-        /// <summary>Returns the hash code for this instance.</summary>
-        public override int GetHashCode() => _value;
-
-        /// <summary>Converts integer number to an Index.</summary>
-        [SuppressMessage("Usage", "CA2225:Operator overloads have named alternates", Justification = "Matches implementation from dotnet/runtime")]
-        public static implicit operator Index(int value) => FromStart(value);
-
-        /// <summary>Converts the value of the current Index object to its equivalent string representation.</summary>
-        [SuppressMessage("Globalization", "CA1305:Specify IFormatProvider", Justification = "Matches implementation from dotnet/runtime")]
-        public override string ToString()
-            => IsFromEnd
-                ? '^' + Value.ToString()
-            : ((uint)Value).ToString();
-    }
-
-    /// <summary>Represent a range has start and end indexes.</summary>
-    /// <remarks>
-    /// Range is used by the C# compiler to support the range syntax.
-    /// <code>
-    /// int[] someArray = new int[5] { 1, 2, 3, 4, 5 };
-    /// int[] subArray1 = someArray[0..2]; // { 1, 2 }
-    /// int[] subArray2 = someArray[1..^0]; // { 2, 3, 4, 5 }
-    /// </code>
-    /// </remarks>
-    [SuppressMessage("Performance", "CA1815:Override equals and operator equals on value types", Justification = "Matches implementation from dotnet/runtime")]
-    [SuppressMessage("Usage", "CA2231:Overload operator equals on overriding value type Equals", Justification = "Matches implementation from dotnet/runtime")]
-    internal readonly struct Range : IEquatable<Range>
-    {
-        /// <summary>Gets represent the inclusive start index of the Range.</summary>
-        public Index Start { get; }
-
-        /// <summary>Gets represent the exclusive end index of the Range.</summary>
-        public Index End { get; }
-
-        /// <summary>Initializes a new instance of the <see cref="Range"/> struct.Construct a Range object using the start and end indexes.</summary>
-        /// <param name="start">Represent the inclusive start index of the range.</param>
-        /// <param name="end">Represent the exclusive end index of the range.</param>
-        public Range(Index start, Index end)
-        {
-            Start = start;
-            End = end;
-        }
-
-        /// <summary>Indicates whether the current Range object is equal to another object of the same type.</summary>
-        /// <param name="obj">An object to compare with this object.</param>
-        public override bool Equals(object? obj) =>
-            obj is Range r &&
-            r.Start.Equals(Start) &&
-            r.End.Equals(End);
-
-        /// <summary>Indicates whether the current Range object is equal to another Range object.</summary>
-        /// <param name="other">An object to compare with this object.</param>
-        public bool Equals(Range other) => other.Start.Equals(Start) && other.End.Equals(End);
-
-        /// <summary>Returns the hash code for this instance.</summary>
-        public override int GetHashCode()
-            => RoslynHashCode.Combine(Start.GetHashCode(), End.GetHashCode());
-
-        /// <summary>Converts the value of the current Range object to its equivalent string representation.</summary>
-        public override string ToString()
-            => $"{Start}..{End}";
-
-        /// <summary>Create a Range object starting from start index to the end of the collection.</summary>
-        public static Range StartAt(Index start) => new(start, Index.End);
-
-        /// <summary>Create a Range object starting from first element in the collection to the end Index.</summary>
-        public static Range EndAt(Index end) => new(Index.Start, end);
-
-        /// <summary>Gets create a Range object starting from first element to the end.</summary>
-        public static Range All => new(Index.Start, Index.End);
-
-        /// <summary>Calculate the start offset and length of range object using a collection length.</summary>
-        /// <param name="length">The length of the collection that the range will be used with. length has to be a positive value.</param>
-        /// <remarks>
-        /// For performance reason, we don't validate the input length parameter against negative values.
-        /// It is expected Range will be used with collections which always have non negative length/count.
-        /// We validate the range is inside the length scope though.
-        /// </remarks>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public (int Offset, int Length) GetOffsetAndLength(int length)
-        {
-            int start;
-            Index startIndex = Start;
-            start = startIndex.IsFromEnd
-                ? length - startIndex.Value
-                : startIndex.Value;
-
-            int end;
-            Index endIndex = End;
-            end = endIndex.IsFromEnd
-                ? length - endIndex.Value
-                : endIndex.Value;
-
-            return (uint)end > (uint)length || (uint)start > (uint)end
-                ? throw new ArgumentOutOfRangeException(nameof(length))
-                : (start, end - start);
-        }
-    }
-#endif
-}
-
-// This was copied from https://github.com/dotnet/coreclr/blob/60f1e6265bd1039f023a82e0643b524d6aaf7845/src/System.Private.CoreLib/shared/System/Diagnostics/CodeAnalysis/NullableAttributes.cs
-// and updated to have the scope of the attributes be internal.
-namespace System.Diagnostics.CodeAnalysis
-{
-    /// <summary>Specifies that null is allowed as an input even if the corresponding type disallows it.</summary>
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Parameter | AttributeTargets.Property, Inherited = false)]
-    internal sealed class AllowNullAttribute : Attribute { }
-#pragma warning restore SA1502 // Element should not be on a single line
-
-    /// <summary>Specifies that null is disallowed as an input even if the corresponding type allows it.</summary>
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Parameter | AttributeTargets.Property, Inherited = false)]
-    internal sealed class DisallowNullAttribute : Attribute
-    {
-    }
-
-    /// <summary>Specifies that an output may be null even if the corresponding type disallows it.</summary>
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Parameter | AttributeTargets.Property | AttributeTargets.ReturnValue, Inherited = false)]
-    internal sealed class MaybeNullAttribute : Attribute
-    {
-    }
-
-    /// <summary>Specifies that an output will not be null even if the corresponding type allows it.</summary>
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Parameter | AttributeTargets.Property | AttributeTargets.ReturnValue, Inherited = false)]
-    internal sealed class NotNullAttribute : Attribute
-    {
-    }
-
-    /// <summary>Specifies that when a method returns <see cref="ReturnValue"/>, the parameter may be null even if the corresponding type disallows it.</summary>
-    [AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
-    internal sealed class MaybeNullWhenAttribute : Attribute
-    {
-        /// <summary>Initializes a new instance of the <see cref="MaybeNullWhenAttribute"/> class.Initializes the attribute with the specified return value condition.</summary>
-        /// <param name="returnValue">
-        /// The return value condition. If the method returns this value, the associated parameter may be null.
-        /// </param>
-        public MaybeNullWhenAttribute(bool returnValue) => ReturnValue = returnValue;
-
-        /// <summary>Gets a value indicating whether gets the return value condition.</summary>
-        public bool ReturnValue { get; }
-    }
-
-    /// <summary>Specifies that when a method returns <see cref="ReturnValue"/>, the parameter will not be null even if the corresponding type allows it.</summary>
-    [AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
-    internal sealed class NotNullWhenAttribute : Attribute
-    {
-        /// <summary>Initializes a new instance of the <see cref="NotNullWhenAttribute"/> class.Initializes the attribute with the specified return value condition.</summary>
-        /// <param name="returnValue">
-        /// The return value condition. If the method returns this value, the associated parameter will not be null.
-        /// </param>
-        public NotNullWhenAttribute(bool returnValue) => ReturnValue = returnValue;
-
-        /// <summary>Gets a value indicating whether gets the return value condition.</summary>
-        public bool ReturnValue { get; }
-    }
-
-    /// <summary>Specifies that the output will be non-null if the named parameter is non-null.</summary>
-    [AttributeUsage(AttributeTargets.Parameter | AttributeTargets.Property | AttributeTargets.ReturnValue, AllowMultiple = true, Inherited = false)]
-    internal sealed class NotNullIfNotNullAttribute : Attribute
-    {
-        /// <summary>Initializes a new instance of the <see cref="NotNullIfNotNullAttribute"/> class.Initializes the attribute with the associated parameter name.</summary>
-        /// <param name="parameterName">
-        /// The associated parameter name.  The output will be non-null if the argument to the parameter specified is non-null.
-        /// </param>
-        public NotNullIfNotNullAttribute(string parameterName) => ParameterName = parameterName;
-
-        /// <summary>Gets the associated parameter name.</summary>
-        public string ParameterName { get; }
-    }
-
-    /// <summary>Applied to a method that will never return under any circumstance.</summary>
-    [AttributeUsage(AttributeTargets.Method, Inherited = false)]
-    internal sealed class DoesNotReturnAttribute : Attribute
-    {
-    }
-
-    /// <summary>Specifies that the method will not return if the associated Boolean parameter is passed the specified value.</summary>
-    [AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
-    internal sealed class DoesNotReturnIfAttribute : Attribute
-    {
-        /// <summary>Initializes a new instance of the <see cref="DoesNotReturnIfAttribute"/> class.Initializes the attribute with the specified parameter value.</summary>
-        /// <param name="parameterValue">
-        /// The condition parameter value. Code after the method will be considered unreachable by diagnostics if the argument to
-        /// the associated parameter matches this value.
-        /// </param>
-        public DoesNotReturnIfAttribute(bool parameterValue) => ParameterValue = parameterValue;
-
-        /// <summary>Gets a value indicating whether gets the condition parameter value.</summary>
-        public bool ParameterValue { get; }
-    }
-
-    /// <summary>Specifies that the method or property will ensure that the listed field and property members have not-null values.</summary>
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, Inherited = false, AllowMultiple = true)]
-    internal sealed class MemberNotNullAttribute : Attribute
-    {
-        /// <summary>Initializes a new instance of the <see cref="MemberNotNullAttribute"/> class.Initializes the attribute with a field or property member.</summary>
-        /// <param name="member">
-        /// The field or property member that is promised to be not-null.
-        /// </param>
-        public MemberNotNullAttribute(string member) => Members = [member];
-
-        /// <summary>Initializes a new instance of the <see cref="MemberNotNullAttribute"/> class.Initializes the attribute with the list of field and property members.</summary>
-        /// <param name="members">
-        /// The list of field and property members that are promised to be not-null.
-        /// </param>
-        public MemberNotNullAttribute(params string[] members) => Members = members;
-
-        /// <summary>Gets field or property member names.</summary>
-        public string[] Members { get; }
-    }
-
-    /// <summary>Specifies that the method or property will ensure that the listed field and property members have not-null values when returning with the specified return value condition.</summary>
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, Inherited = false, AllowMultiple = true)]
-    internal sealed class MemberNotNullWhenAttribute : Attribute
-    {
-        /// <summary>Initializes a new instance of the <see cref="MemberNotNullWhenAttribute"/> class.Initializes the attribute with the specified return value condition and a field or property member.</summary>
-        /// <param name="returnValue">
-        /// The return value condition. If the method returns this value, the associated parameter will not be null.
-        /// </param>
-        /// <param name="member">
-        /// The field or property member that is promised to be not-null.
-        /// </param>
-        public MemberNotNullWhenAttribute(bool returnValue, string member)
-        {
-            ReturnValue = returnValue;
-            Members = [member];
-        }
-
-        /// <summary>Initializes a new instance of the <see cref="MemberNotNullWhenAttribute"/> class.Initializes the attribute with the specified return value condition and list of field and property members.</summary>
-        /// <param name="returnValue">
-        /// The return value condition. If the method returns this value, the associated parameter will not be null.
-        /// </param>
-        /// <param name="members">
-        /// The list of field and property members that are promised to be not-null.
-        /// </param>
-        public MemberNotNullWhenAttribute(bool returnValue, params string[] members)
-        {
-            ReturnValue = returnValue;
-            Members = members;
-        }
-
-        /// <summary>Gets a value indicating whether gets the return value condition.</summary>
-        public bool ReturnValue { get; }
-
-        /// <summary>Gets field or property member names.</summary>
-        public string[] Members { get; }
-    }
-}
-
-namespace System.Text
-{
-    internal static class StringBuilderExtensions
-    {
-        public static StringBuilder Append(this StringBuilder builder, IFormatProvider? _, string value)
-            => builder.Append(value);
-
-        public static StringBuilder AppendLine(this StringBuilder builder, IFormatProvider? _, string value)
-            => builder.AppendLine(value);
-    }
-}
-
-#endif
-
-#if NETCOREAPP3_1
-namespace System.Diagnostics.CodeAnalysis
-{
-    /// <summary>Specifies that the method or property will ensure that the listed field and property members have not-null values.</summary>
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, Inherited = false, AllowMultiple = true)]
-    internal sealed class MemberNotNullAttribute : Attribute
-    {
-        /// <summary>Initializes the attribute with a field or property member.</summary>
-        /// <param name="member">
-        /// The field or property member that is promised to be not-null.
-        /// </param>
-        public MemberNotNullAttribute(string member) => Members = [member];
-
-        /// <summary>Initializes the attribute with the list of field and property members.</summary>
-        /// <param name="members">
-        /// The list of field and property members that are promised to be not-null.
-        /// </param>
-        public MemberNotNullAttribute(params string[] members) => Members = members;
-
-        /// <summary>Gets field or property member names.</summary>
-        public string[] Members { get; }
-    }
-
-    /// <summary>Specifies that the method or property will ensure that the listed field and property members have not-null values when returning with the specified return value condition.</summary>
-    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, Inherited = false, AllowMultiple = true)]
-    internal sealed class MemberNotNullWhenAttribute : Attribute
-    {
-        /// <summary>Initializes the attribute with the specified return value condition and a field or property member.</summary>
-        /// <param name="returnValue">
-        /// The return value condition. If the method returns this value, the associated parameter will not be null.
-        /// </param>
-        /// <param name="member">
-        /// The field or property member that is promised to be not-null.
-        /// </param>
-        public MemberNotNullWhenAttribute(bool returnValue, string member)
-        {
-            ReturnValue = returnValue;
-            Members = [member];
-        }
-
-        /// <summary>Initializes the attribute with the specified return value condition and list of field and property members.</summary>
-        /// <param name="returnValue">
-        /// The return value condition. If the method returns this value, the associated parameter will not be null.
-        /// </param>
-        /// <param name="members">
-        /// The list of field and property members that are promised to be not-null.
-        /// </param>
-        public MemberNotNullWhenAttribute(bool returnValue, params string[] members)
-        {
-            ReturnValue = returnValue;
-            Members = members;
-        }
-
-        /// <summary>Gets the return value condition.</summary>
-        public bool ReturnValue { get; }
-
-        /// <summary>Gets field or property member names.</summary>
-        public string[] Members { get; }
-    }
-}
-#endif
-
-#if !NET7_0_OR_GREATER
-namespace System.Diagnostics.CodeAnalysis
-{
-    /// <summary>
-    /// Specifies that this constructor sets all required members for the current type,
-    /// and callers do not need to set any required members themselves.
-    /// </summary>
-    [AttributeUsage(AttributeTargets.Constructor, AllowMultiple = false, Inherited = false)]
-    internal sealed class SetsRequiredMembersAttribute : Attribute
-    {
-    }
-}
-
-#pragma warning disable SA1403 // File may only contain a single namespace
-namespace System.Runtime.CompilerServices
-{
-    /// <summary>
-    /// Indicates that compiler support for a particular feature is required for the location where this attribute is applied.
-    /// </summary>
-    [AttributeUsage(AttributeTargets.All, AllowMultiple = true, Inherited = false)]
-    internal sealed class CompilerFeatureRequiredAttribute : Attribute
-    {
-        /// <summary>
-        /// The <see cref="FeatureName"/> used for the ref structs C# feature.
-        /// </summary>
-        public const string RefStructs = nameof(RefStructs);
-
-        /// <summary>
-        /// The <see cref="FeatureName"/> used for the required members C# feature.
-        /// </summary>
-        public const string RequiredMembers = nameof(RequiredMembers);
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="CompilerFeatureRequiredAttribute"/> class.
-        /// </summary>
-        /// <param name="featureName">The name of the feature to indicate.</param>
-        public CompilerFeatureRequiredAttribute(string featureName)
-        {
-            FeatureName = featureName;
-        }
-
-        /// <summary>
-        /// Gets the name of the compiler feature.
-        /// </summary>
-        public string FeatureName { get; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether if true, the compiler can choose to allow access to the location where this attribute is applied if it does not understand <see cref="FeatureName"/>.
-        /// </summary>
-        public bool IsOptional { get; set; }
-    }
-
-    /// <summary>
-    /// Specifies that a type has required members or that a member is required.
-    /// </summary>
-    [AttributeUsage(
-        AttributeTargets.Class |
-        AttributeTargets.Struct |
-        AttributeTargets.Field |
-        AttributeTargets.Property,
-        AllowMultiple = false,
-        Inherited = false)]
-    internal sealed class RequiredMemberAttribute : Attribute
-    {
-    }
-}
-#endif
